fix: change return type of SafeDiv function from INT to FLOAT

+++ MySQL_Advanced/10-div.sql
@@ -1,14 +1,14 @@
 DELIMITER $$
--- function :)
+
 CREATE FUNCTION SafeDiv (a INT, b INT)
-RETURNS INT
+RETURNS FLOAT
 DETERMINISTIC
 BEGIN
-   IF b = 0 THEN
-      RETURN 0;
-   END IF;
+    IF b = 0 THEN
+    RETURN 0;
+    END IF;
 
-   RETURN a / b;
+    RETURN a / b;
 END$$
 
 DELIMITER ;
